Turn admin view into normal view
