+++ app/grandchallenge/evaluation/admin.py
@@ -1,22 +1,16 @@
 import json
 
-from django.contrib import admin, messages
-from django.contrib.auth.mixins import PermissionRequiredMixin
+from django.contrib import admin
 from django.core.exceptions import ValidationError
-from django.forms import Form, ModelForm, ModelMultipleChoiceField
-from django.urls import path, reverse
+from django.forms import ModelForm
 from django.utils.html import format_html
-from django.utils.text import slugify
-from django.views.generic import FormView
 
-from grandchallenge.archives.models import Archive
 from grandchallenge.components.admin import (
     ComponentImageAdmin,
     cancel_jobs,
     deprovision_jobs,
     requeue_jobs,
 )
-from grandchallenge.components.models import ComponentInterface
 from grandchallenge.core.admin import (
     GroupObjectPermissionAdmin,
     UserObjectPermissionAdmin,
@@ -41,7 +35,6 @@ from grandchallenge.evaluation.models import (
     SubmissionUserObjectPermission,
 )
 from grandchallenge.evaluation.tasks import create_evaluation
-from grandchallenge.evaluation.utils import SubmissionKindChoices
 
 
 class PhaseAdminForm(ModelForm):
@@ -63,84 +56,6 @@ class PhaseAdminForm(ModelForm):
             )
 
         return cleaned_data
-
-
-class ConfigureAlgorithmPhasesForm(Form):
-    phases = ModelMultipleChoiceField(
-        queryset=Phase.objects.select_related("challenge")
-        .filter(submission_kind=SubmissionKindChoices.CSV)
-        .all()
-    )
-    algorithm_inputs = ModelMultipleChoiceField(
-        queryset=ComponentInterface.objects.all()
-    )
-    algorithm_outputs = ModelMultipleChoiceField(
-        queryset=ComponentInterface.objects.all()
-    )
-
-    def clean(self):
-        cleaned_data = super().clean()
-        try:
-            for phase in self.cleaned_data["phases"]:
-                if Archive.objects.filter(
-                    slug=f"{slugify(phase.challenge.short_name)}-{slugify(phase.title)}-dataset"
-                ).exists():
-                    raise ValidationError(
-                        f"Archive for {phase} already exists."
-                    )
-        except KeyError:
-            pass
-        return cleaned_data
-
-
-class ConfigureAlgorithmPhasesView(PermissionRequiredMixin, FormView):
-    form_class = ConfigureAlgorithmPhasesForm
-    permission_required = "evaluation.configure_algorithm_phase"
-    template_name = "admin/evaluation/phase/algorithm_phase_form.html"
-    raise_exception = True
-
-    def form_valid(self, form):
-        for phase in form.cleaned_data["phases"]:
-            self.turn_phase_into_algorithm_phase(
-                phase=phase,
-                inputs=form.cleaned_data["algorithm_inputs"],
-                outputs=form.cleaned_data["algorithm_outputs"],
-            )
-        messages.success(self.request, "Phases were successfully updated")
-        return super().form_valid(form)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context.update(
-            {
-                **admin.site.each_context(self.request),
-                "opts": Phase._meta,
-                "title": "Turn phase(s) into algorithm phase(s)",
-            }
-        )
-        return context
-
-    def get_success_url(self):
-        return reverse("admin:evaluation_phase_changelist")
-
-    def turn_phase_into_algorithm_phase(self, *, phase, inputs, outputs):
-        archive = Archive.objects.create(
-            title=format_html(
-                "{challenge_name} {phase_title} dataset",
-                challenge_name=phase.challenge.short_name,
-                phase_title=phase.title,
-            )
-        )
-
-        for user in phase.challenge.admins_group.user_set.all():
-            archive.add_editor(user)
-
-        phase.archive = archive
-        phase.submission_kind = phase.SubmissionKindChoices.ALGORITHM
-        phase.creator_must_be_verified = True
-        phase.save()
-        phase.algorithm_outputs.add(*outputs)
-        phase.algorithm_inputs.add(*inputs)
 
 
 @admin.register(Phase)
@@ -174,20 +89,6 @@ class PhaseAdmin(admin.ModelAdmin):
         "challenge_forge_json",
     )
     form = PhaseAdminForm
-    change_list_template = "admin/evaluation/phase/change_list.html"
-
-    def get_urls(self):
-        urls = super().get_urls()
-        extra_url = [
-            path(
-                "create_algorithm_phase/",
-                self.admin_site.admin_view(
-                    ConfigureAlgorithmPhasesView.as_view()
-                ),
-                name="algorithm_phase_create",
-            )
-        ]
-        return extra_url + urls
 
     @admin.display(boolean=True)
     def open_for_submissions(self, instance):
