--- app/grandchallenge/evaluation/templates/admin/evaluation/phase/algorithm_phase_form.html
@@ -1,9 +1,0 @@
-{% extends "admin/change_form.html" %}
-
-{% block content %}
-    <form method="post">
-        {% csrf_token %}
-        {{ form.as_p }}
-        <input type="submit" value="Submit">
-    </form>
-{% endblock %}
