--- app/grandchallenge/evaluation/templates/admin/evaluation/phase/change_list.html
@@ -1,8 +1,0 @@
-{% extends "admin/change_list.html" %}
-
-{% block object-tools-items %}
-    {{ block.super }}
-    <li>
-        <a href="{% url 'admin:algorithm_phase_create' %}">Configure algorithm phases</a>
-    </li>
-{% endblock %}
