+++ app/grandchallenge/evaluation/views/__init__.py
@@ -4,12 +4,14 @@ from dateutil.relativedelta import relativedelta
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth import get_user_model
+from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.contrib.messages.views import SuccessMessageMixin
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import Q
 from django.http import Http404
 from django.shortcuts import get_object_or_404
 from django.utils.functional import cached_property
+from django.utils.html import format_html
 from django.utils.timezone import now
 from django.views.generic import (
     CreateView,
@@ -24,7 +26,10 @@ from guardian.mixins import LoginRequiredMixin
 
 from grandchallenge.algorithms.forms import AlgorithmForPhaseForm
 from grandchallenge.algorithms.models import Algorithm, Job
+from grandchallenge.archives.models import Archive
+from grandchallenge.challenges.models import ChallengeRequest
 from grandchallenge.components.models import InterfaceKind
+from grandchallenge.core.fixtures import create_uploaded_image
 from grandchallenge.core.forms import UserFormKwargsMixin
 from grandchallenge.core.guardian import (
     ObjectPermissionRequiredMixin,
@@ -35,6 +40,7 @@ from grandchallenge.datatables.views import Column, PaginatedTableListView
 from grandchallenge.direct_messages.forms import ConversationForm
 from grandchallenge.evaluation.forms import (
     CombinedLeaderboardForm,
+    ConfigureAlgorithmPhasesForm,
     EvaluationForm,
     MethodForm,
     MethodUpdateForm,
@@ -54,6 +60,7 @@ from grandchallenge.evaluation.utils import SubmissionKindChoices
 from grandchallenge.subdomains.utils import reverse, reverse_lazy
 from grandchallenge.teams.models import Team
 from grandchallenge.verifications.views import VerificationRequiredMixin
+from grandchallenge.workstations.models import Workstation
 
 
 class CachedPhaseMixin:
@@ -980,3 +987,65 @@ class CombinedLeaderboardDelete(
 
     def get_permission_object(self):
         return self.get_object().challenge
+
+
+class ConfigureAlgorithmPhasesView(PermissionRequiredMixin, FormView):
+    form_class = ConfigureAlgorithmPhasesForm
+    permission_required = "evaluation.configure_algorithm_phase"
+    template_name = "evaluation/configure_algorithm_phases_form.html"
+    raise_exception = True
+
+    def form_valid(self, form):
+        for phase in form.cleaned_data["phases"]:
+            self.turn_phase_into_algorithm_phase(
+                phase=phase,
+                inputs=form.cleaned_data["algorithm_inputs"],
+                outputs=form.cleaned_data["algorithm_outputs"],
+            )
+        messages.success(self.request, "Phases were successfully updated")
+        return super().form_valid(form)
+
+    def get_success_url(self):
+        return reverse("challenges:list")
+
+    def turn_phase_into_algorithm_phase(self, *, phase, inputs, outputs):
+        archive = Archive.objects.create(
+            title=format_html(
+                "{challenge_name} {phase_title} dataset",
+                challenge_name=phase.challenge.short_name,
+                phase_title=phase.title,
+            ),
+            workstation=Workstation.objects.get(
+                slug=settings.DEFAULT_WORKSTATION_SLUG
+            ),
+            logo=(
+                phase.challenge.logo
+                if phase.challenge.logo
+                else create_uploaded_image()
+            ),
+        )
+        archive.full_clean()
+        archive.save()
+
+        for user in phase.challenge.admins_group.user_set.all():
+            archive.add_editor(user)
+
+        try:
+            challenge_request = ChallengeRequest.objects.get(
+                short_name=phase.challenge.short_name
+            )
+            phase.algorithm_time_limit = (
+                challenge_request.inference_time_limit_in_minutes
+            )
+        except ObjectDoesNotExist:
+            messages.warning(
+                self.request,
+                f"Algorithm run time limit could not be configured for {phase}.",
+            )
+
+        phase.archive = archive
+        phase.submission_kind = phase.SubmissionKindChoices.ALGORITHM
+        phase.creator_must_be_verified = True
+        phase.save()
+        phase.algorithm_outputs.add(*outputs)
+        phase.algorithm_inputs.add(*inputs)
