+++ app/tests/evaluation_tests/test_forms.py
@@ -3,7 +3,10 @@ from factory.django import ImageField
 
 from grandchallenge.algorithms.forms import AlgorithmForPhaseForm
 from grandchallenge.algorithms.models import AlgorithmImage
-from grandchallenge.evaluation.forms import SubmissionForm
+from grandchallenge.evaluation.forms import (
+    ConfigureAlgorithmPhasesForm,
+    SubmissionForm,
+)
 from grandchallenge.evaluation.models import Phase, Submission
 from grandchallenge.evaluation.utils import SubmissionKindChoices
 from grandchallenge.invoices.models import PaymentStatusChoices
@@ -20,6 +23,7 @@ from tests.components_tests.factories import (
 )
 from tests.evaluation_tests.factories import (
     EvaluationFactory,
+    MethodFactory,
     PhaseFactory,
     SubmissionFactory,
 )
@@ -529,3 +533,33 @@ def test_algorithm_for_phase_form_validation():
         "You have already created the maximum number of algorithms for this phase."
         in str(form.errors)
     )
+
+
+@pytest.mark.django_db
+def test_configure_algorithm_phases_form():
+    p1, p2, p3 = PhaseFactory.create_batch(
+        3, submission_kind=SubmissionKindChoices.CSV
+    )
+    SubmissionFactory(phase=p1)
+    MethodFactory(phase=p2)
+    p_alg = PhaseFactory(submission_kind=SubmissionKindChoices.ALGORITHM)
+    ci1, ci2 = ComponentInterfaceFactory.create_batch(2)
+
+    form = ConfigureAlgorithmPhasesForm(
+        data={
+            "phases": [p1, p2, p3, p_alg],
+            "algorithm_inputs": [ci1],
+            "algorithm_outputs": [ci2],
+        }
+    )
+    assert form.is_valid() is False
+    assert list(form.fields["phases"].queryset) == [p3]
+
+    form3 = ConfigureAlgorithmPhasesForm(
+        data={
+            "phases": [p3],
+            "algorithm_inputs": [ci1],
+            "algorithm_outputs": [ci2],
+        }
+    )
+    assert form3.is_valid()
